+++ bundles/framework/findbycoordinates/resources/locale/fi.js
@@ -9,10 +9,15 @@ Oskari.registerLocalization(
             "tooltip": "Klikkaa pistettä kartalla. Hakutuloksena saat pisteen koordinaatit ja pistettä lähimmän osoitteen."
         },
         "resultsTitle": "Hakutulokset",
-        "grid": {
-            "address": "Osoite",
-            "coordinates": "Koordinaatit"
-        }
+        "close": "Sulje",
+        "channels" : {
+            "WHAT3WORDS_CHANNEL": "what3words (FI)",
+            "NLS_NEAREST_FEATURE_CHANNEL": "Lähin osoite"
+        },
+        "channelDescriptions" : {
+            "WHAT3WORDS_CHANNEL": "what3words on 3 sanan yhdistelmä, joka määrittää tietyn 3 m x 3 m -ruudun missä tahansa maapallolla."
+        },
+        "popupTitle": "Hakutulokset"
     }
 }
 );